The Hiring Process
The process is the same for all entry-level US post office jobs, It is online.
No experience is required, and all postal hiring decisions are based on merit and how well you compete in this post office hiring process, not how well you have done in the past. USPS Jobs is an Equal Opportunity Employer.